fix: format, division indicators file
