+++ data/src/chart/indicator/config.rs
@@ -1,125 +1,9 @@
-use std::fmt::{self, Debug, Display};
+use std::fmt::{self, Display};
 
-use enum_map::Enum;
-use exchange::adapter::MarketKind;
 use serde::{Deserialize, Serialize};
 
+use super::core::{HeatmapIndicator, KlineIndicator};
 use crate::chart::style::BuySellColors;
-
-pub trait Indicator: PartialEq + Display + 'static {
-    fn for_market(market: MarketKind) -> &'static [Self]
-    where
-        Self: Sized;
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Deserialize, Serialize, Eq, Enum)]
-pub enum KlineIndicator {
-    Volume,
-    BarAnalysis,
-    CumulativeDelta,
-    OpenInterest,
-}
-
-impl Indicator for KlineIndicator {
-    fn for_market(market: MarketKind) -> &'static [Self] {
-        match market {
-            MarketKind::Spot => &Self::FOR_SPOT,
-            MarketKind::LinearPerps | MarketKind::InversePerps => &Self::FOR_PERPS,
-        }
-    }
-}
-
-impl KlineIndicator {
-    // Indicator togglers on UI menus depend on these arrays.
-    // Every variant needs to be in either SPOT, PERPS or both.
-    /// Indicators that can be used with spot market tickers
-    const FOR_SPOT: [KlineIndicator; 3] = [
-        KlineIndicator::Volume,
-        KlineIndicator::BarAnalysis,
-        KlineIndicator::CumulativeDelta,
-    ];
-    /// Indicators that can be used with perpetual swap market tickers
-    const FOR_PERPS: [KlineIndicator; 4] = [
-        KlineIndicator::Volume,
-        KlineIndicator::BarAnalysis,
-        KlineIndicator::CumulativeDelta,
-        KlineIndicator::OpenInterest,
-    ];
-}
-
-impl Display for KlineIndicator {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            KlineIndicator::Volume => write!(f, "Volume"),
-            KlineIndicator::BarAnalysis => write!(f, "Bar Analysis"),
-            KlineIndicator::CumulativeDelta => write!(f, "CVD"),
-            KlineIndicator::OpenInterest => write!(f, "Open Interest"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Deserialize, Serialize, Eq, Enum)]
-pub enum HeatmapIndicator {
-    Volume,
-}
-
-impl Indicator for HeatmapIndicator {
-    fn for_market(market: MarketKind) -> &'static [Self] {
-        match market {
-            MarketKind::Spot => &Self::FOR_SPOT,
-            MarketKind::LinearPerps | MarketKind::InversePerps => &Self::FOR_PERPS,
-        }
-    }
-}
-
-impl HeatmapIndicator {
-    // Indicator togglers on UI menus depend on these arrays.
-    // Every variant needs to be in either SPOT, PERPS or both.
-    /// Indicators that can be used with spot market tickers
-    const FOR_SPOT: [HeatmapIndicator; 1] = [HeatmapIndicator::Volume];
-    /// Indicators that can be used with perpetual swap market tickers
-    const FOR_PERPS: [HeatmapIndicator; 1] = [HeatmapIndicator::Volume];
-}
-
-impl Display for HeatmapIndicator {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            HeatmapIndicator::Volume => write!(f, "Volume"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-/// Temporary workaround,
-/// represents any indicator type in the UI
-pub enum UiIndicator {
-    Heatmap(HeatmapIndicator),
-    Kline(KlineIndicator),
-}
-
-impl From<KlineIndicator> for UiIndicator {
-    fn from(k: KlineIndicator) -> Self {
-        UiIndicator::Kline(k)
-    }
-}
-
-impl From<HeatmapIndicator> for UiIndicator {
-    fn from(h: HeatmapIndicator) -> Self {
-        UiIndicator::Heatmap(h)
-    }
-}
-
-impl PartialEq for UiIndicator {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (UiIndicator::Heatmap(a), UiIndicator::Heatmap(b)) => a == b,
-            (UiIndicator::Kline(a), UiIndicator::Kline(b)) => a == b,
-            _ => false,
-        }
-    }
-}
-
-impl Eq for UiIndicator {}
 
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 #[serde(default)]
